apps/api/requirements.txt
fastapi>=0.116.2
uvicorn[standard]==0.24.0
pydantic>=2.5.0
cognee==0.5.4
langchain>=0.2.0
neo4j==5.14.1
python-dotenv>=1.0.1
aiofiles==23.2.1
